requirements.txt
fastapi
uvicorn
pyodbc
python-dotenv
gunicorn
pydantic
sqlalchemy